backend_django/pyproject.toml
[tool.poetry]
name = "kdt-final-django"
version = "0.1.0"
description = ""
authors = ["SonJinHYo <wlsgy95@naver.com>"]
license = "\"MIT\""
readme = "README.md"
packages = [{include = "kdt_final_django"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-environ = "^0.9.0"
requests = "^2.28.2"
django-cloudflare-images = "^0.4.1"
transformers = "^4.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
